(File) rename textbox.tsx to teatarea.tsx
